+++ tests/notebooks/R/simple_r_script.R
@@ -1,0 +1,5 @@
+# This is a comment
+
+cars
+
+plot(cars)