+++ tests/test_active_cells.py
@@ -11,11 +11,11 @@ HEADER = {'.py': '''# ---
 # ---
 
 ''',
-          '.R': '''#' ---
-#' jupyter:
-#'   jupytext:
-#'     main_language: python
-#' ---
+          '.R': '''# ---
+# jupyter:
+#   jupytext:
+#     main_language: python
+# ---
 
 ''',
 
@@ -34,7 +34,7 @@ ACTIVE_ALL = {'.py': """# + {"active": "ipynb,py,R,Rmd"}
 # This cell is active in all extensions
 ```
 """,
-              '.R': """#+ active="ipynb,py,R,Rmd"
+              '.R': """# + {"active": "ipynb,py,R,Rmd"}
 # This cell is active in all extensions
 """,
               '.ipynb': {'cell_type': 'code',
@@ -47,10 +47,10 @@ ACTIVE_ALL = {'.py': """# + {"active": "ipynb,py,R,Rmd"}
 @pytest.mark.parametrize('ext', ['.Rmd', '.py', '.R'])
 def test_active_all(ext):
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', False):
-        nb = jupytext.reads(ACTIVE_ALL[ext], ext)
+        nb = jupytext.reads(HEADER[ext] + ACTIVE_ALL[ext], ext)
         assert len(nb.cells) == 1
-        compare(nb.cells[0], ACTIVE_ALL['.ipynb'])
         compare(ACTIVE_ALL[ext], jupytext.writes(nb, ext))
+        compare(nb.cells[0], ACTIVE_ALL['.ipynb'])
 
 
 ACTIVE_IPYNB = {'.py': """# + {"active": "ipynb"}
@@ -62,7 +62,7 @@ ACTIVE_IPYNB = {'.py': """# + {"active": "ipynb"}
 %matplotlib inline
 ```
 """,
-                '.R': """#+ language="python", active="ipynb", eval=FALSE
+                '.R': """# + {"active": "ipynb"}
 # # This cell is active only in ipynb
 # %matplotlib inline
 """,
@@ -80,9 +80,8 @@ def test_active_ipynb(ext):
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', False):
         nb = jupytext.reads(HEADER[ext] + ACTIVE_IPYNB[ext], ext)
         assert len(nb.cells) == 1
+        compare(ACTIVE_IPYNB[ext], jupytext.writes(nb, ext))
         compare(nb.cells[0], ACTIVE_IPYNB['.ipynb'])
-        if ext != '.R':
-            compare(ACTIVE_IPYNB[ext], jupytext.writes(nb, ext))
 
 
 ACTIVE_PY_IPYNB = {'.py': """# + {"active": "ipynb,py"}
@@ -92,7 +91,7 @@ ACTIVE_PY_IPYNB = {'.py': """# + {"active": "ipynb,py"}
 # This cell is active in py and ipynb extensions
 ```
 """,
-                   '.R': """#+ language="python", active="ipynb,py", eval=FALSE
+                   '.R': """# + {"active": "ipynb,py"}
 # # This cell is active in py and ipynb extensions
 """,
                    '.ipynb': {'cell_type': 'code',
@@ -109,9 +108,8 @@ def test_active_py_ipynb(ext):
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', False):
         nb = jupytext.reads(HEADER[ext] + ACTIVE_PY_IPYNB[ext], ext)
         assert len(nb.cells) == 1
+        compare(ACTIVE_PY_IPYNB[ext], jupytext.writes(nb, ext))
         compare(nb.cells[0], ACTIVE_PY_IPYNB['.ipynb'])
-        if ext != '.R':
-            compare(ACTIVE_PY_IPYNB[ext], jupytext.writes(nb, ext))
 
 
 ACTIVE_PY_R_IPYNB = {'.py': """# + {"active": "ipynb,py,R"}
@@ -121,7 +119,7 @@ ACTIVE_PY_R_IPYNB = {'.py': """# + {"active": "ipynb,py,R"}
 # This cell is active in py, R and ipynb extensions
 ```
 """,
-                     '.R': """#+ language="python", active="ipynb,py,R", eval=FALSE
+                     '.R': """# + {"active": "ipynb,py,R"}
 # This cell is active in py, R and ipynb extensions
 """,
                      '.ipynb': {'cell_type': 'code',
@@ -138,9 +136,8 @@ def test_active_py_r_ipynb(ext):
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', False):
         nb = jupytext.reads(HEADER[ext] + ACTIVE_PY_R_IPYNB[ext], ext)
         assert len(nb.cells) == 1
+        compare(ACTIVE_PY_R_IPYNB[ext], jupytext.writes(nb, ext))
         compare(nb.cells[0], ACTIVE_PY_R_IPYNB['.ipynb'])
-        if ext != '.R':
-            compare(ACTIVE_PY_R_IPYNB[ext], jupytext.writes(nb, ext))
 
 
 ACTIVE_RMD = {'.py': """# + {"active": "Rmd"}
@@ -150,7 +147,7 @@ ACTIVE_RMD = {'.py': """# + {"active": "Rmd"}
 # This cell is active in Rmd only
 ```
 """,
-              '.R': """#+ language="python", active="Rmd", eval=FALSE
+              '.R': """# + {"active": "Rmd"}
 # # This cell is active in Rmd only
 """,
               '.ipynb': {'cell_type': 'raw',
@@ -164,9 +161,8 @@ def test_active_rmd(ext):
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', False):
         nb = jupytext.reads(HEADER[ext] + ACTIVE_RMD[ext], ext)
         assert len(nb.cells) == 1
+        compare(ACTIVE_RMD[ext], jupytext.writes(nb, ext))
         compare(nb.cells[0], ACTIVE_RMD['.ipynb'])
-        if ext != '.R':
-            compare(ACTIVE_RMD[ext], jupytext.writes(nb, ext))
 
 
 ACTIVE_NOT_INCLUDE_RMD = {'.py': """# + {"hide_output": true, "active": "Rmd"}
@@ -176,7 +172,7 @@ ACTIVE_NOT_INCLUDE_RMD = {'.py': """# + {"hide_output": true, "active": "Rmd"}
 # This cell is active in Rmd only
 ```
 """,
-                          '.R': """#+ include=FALSE, active="Rmd", eval=FALSE
+                          '.R': """# + {"hide_output": true, "active": "Rmd"}
 # # This cell is active in Rmd only
 """,
                           '.ipynb':
@@ -192,5 +188,5 @@ def test_active_not_include_rmd(ext):
     with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', False):
         nb = jupytext.reads(ACTIVE_NOT_INCLUDE_RMD[ext], ext)
         assert len(nb.cells) == 1
-        compare(nb.cells[0], ACTIVE_NOT_INCLUDE_RMD['.ipynb'])
         compare(ACTIVE_NOT_INCLUDE_RMD[ext], jupytext.writes(nb, ext))
+        compare(nb.cells[0], ACTIVE_NOT_INCLUDE_RMD['.ipynb'])