+++ tests/test_read_simple_R.py
@@ -146,3 +146,25 @@ def test_escape_start_pattern(ext, rnb="""#' The code start pattern '#+' can
 1 + 1''')
     rnb2 = jupytext.writes(nb, ext)
     compare(rnb, rnb2)
+
+
+@pytest.mark.parametrize('ext', ['.r', '.R'])
+def test_read_simple_r(ext, text="""# This is a very simple R file
+# I expect to get three cells here.
+#
+# The first one is markdown. The two others
+# are code cells
+
+cars
+
+plot(cars)
+"""):
+    nb = jupytext.reads(text, ext)
+    assert len(nb.cells) == 3
+    assert nb.cells[0].cell_type == 'markdown'
+    assert nb.cells[1].cell_type == 'code'
+    assert nb.cells[2].cell_type == 'code'
+    assert nb.cells[1].source == 'cars'
+    assert nb.cells[2].source == 'plot(cars)'
+    text2 = jupytext.writes(nb, ext)
+    compare(text, text2)